refactor(tests): consolidate _FakeKB into tests/helpers/fake_kb.py (#5557)
Extracts two inline _FakeKB implementations per #5443/#5557:
- tests/helpers/fake_kb.py: MinimalFakeKB (redis() accessor) + FactsFakeKB(MinimalFakeKB, FactsMixin)
- tests/knowledge/test_facts_dedup.py: FactsFakeKB inline class → import from helpers
- tests/test_knowledge_boards.py: _FakeKB inline → MinimalFakeKB(fake_redis) from helpers
